ref(explore): Move toolbar components into own file
As each component will get more complex, move them into separate files to be
worked on independently.
