+++ static/app/views/explore/toolbar/index.tsx
@@ -1,6 +1,8 @@
-import {t} from 'sentry/locale';
-
-import {ToolbarHeading, ToolbarSection} from './styles';
+import {ToolbarGroupBy} from './toolbarGroupBy';
+import {ToolbarLimitTo} from './toolbarLimitTo';
+import {ToolbarResults} from './toolbarResults';
+import {ToolbarSortBy} from './toolbarSortBy';
+import {ToolbarVisualize} from './toolbarVisualize';
 
 interface ExploreToolbarProps {}
 
@@ -13,57 +15,5 @@ export function ExploreToolbar({}: ExploreToolbarProps) {
       <ToolbarLimitTo />
       <ToolbarGroupBy disabled />
     </div>
-  );
-}
-
-interface ToolbarResultsProps {}
-
-function ToolbarResults({}: ToolbarResultsProps) {
-  return (
-    <ToolbarSection>
-      <ToolbarHeading>{t('Results')}</ToolbarHeading>
-    </ToolbarSection>
-  );
-}
-
-interface ToolbarVisualizeProps {}
-
-function ToolbarVisualize({}: ToolbarVisualizeProps) {
-  return (
-    <ToolbarSection>
-      <ToolbarHeading>{t('Visualize')}</ToolbarHeading>
-    </ToolbarSection>
-  );
-}
-
-interface ToolbarSortByProps {}
-
-function ToolbarSortBy({}: ToolbarSortByProps) {
-  return (
-    <ToolbarSection>
-      <ToolbarHeading>{t('Sort By')}</ToolbarHeading>
-    </ToolbarSection>
-  );
-}
-
-interface ToolbarLimitToProps {}
-
-function ToolbarLimitTo({}: ToolbarLimitToProps) {
-  return (
-    <ToolbarSection>
-      <ToolbarHeading>{t('Limit To')}</ToolbarHeading>
-    </ToolbarSection>
-  );
-}
-
-interface ToolbarGroupByProps {
-  disabled?: boolean;
-}
-
-function ToolbarGroupBy({disabled}: ToolbarGroupByProps) {
-  return (
-    <ToolbarSection>
-      <ToolbarHeading disabled={disabled}>{t('Group By')}</ToolbarHeading>
-    </ToolbarSection>
   );
 }
