+++ static/app/views/explore/toolbar/styles.tsx
@@ -1,6 +1,10 @@
 import styled from '@emotion/styled';
 
-export const ToolbarSection = styled('div')``;
+import {space} from 'sentry/styles/space';
+
+export const ToolbarSection = styled('div')`
+  margin-bottom: ${space(2)};
+`;
 
 export const ToolbarHeading = styled('h6')<{disabled?: boolean}>`
   color: ${p => (p.disabled ? p.theme.gray300 : p.theme.purple300)};
@@ -10,4 +14,5 @@ export const ToolbarHeading = styled('h6')<{disabled?: boolean}>`
   line-height: ${p => p.theme.form.md.lineHeight};
   text-decoration: underline dotted
     ${p => (p.disabled ? p.theme.gray300 : p.theme.purple300)};
+  margin: 0 0 ${space(1)} 0;
 `;
