+++ static/app/views/explore/toolbar/toolbarGroupBy.tsx
@@ -1,0 +1,15 @@
+import {t} from 'sentry/locale';
+
+import {ToolbarHeading, ToolbarSection} from './styles';
+
+interface ToolbarGroupByProps {
+  disabled?: boolean;
+}
+
+export function ToolbarGroupBy({disabled}: ToolbarGroupByProps) {
+  return (
+    <ToolbarSection>
+      <ToolbarHeading disabled={disabled}>{t('Group By')}</ToolbarHeading>
+    </ToolbarSection>
+  );
+}
